+++ src/jsonData.py
@@ -1,0 +1,2 @@
+#save the previous ran data into a json file 
+
